admin-dash

+++ app/Listeners/UpdateSessionVariables.php
@@ -25,55 +25,53 @@ class UpdateSessionVariables
 
     public function handle(Login $event)
     {
-        $user = $event->user;
-
-        $cart_ids = Cart::where('session_id', Session::get('customer_unique_id'))->select('id')->get();
-        $wish_ids = Wishlist::where('session_id', Session::get('customer_unique_id'))->select('id')->get();
-
-        foreach($cart_ids as $id)
+        if(!auth()->guard('admin')->check())
         {
-            $session_cart = Cart::find($id['id']);
+            $user = $event->user;
 
-            if($user_cart = Cart::where('user_id', $user->id)
-            ->where('volume_id',$session_cart->volume_id)
-            ->where('attribute_id',$session_cart->attribute_id)->first())
-            {
-                $user_cart->quantity += $session_cart->quantity;
+            $cart_ids = Cart::where('session_id', Session::get('customer_unique_id'))->select('id')->get();
+            $wish_ids = Wishlist::where('session_id', Session::get('customer_unique_id'))->select('id')->get();
 
-                $user_cart->save();
+            foreach ($cart_ids as $id) {
+                $session_cart = Cart::find($id['id']);
 
-                $session_cart->delete();
+                if ($user_cart = Cart::where('user_id', $user->id)
+                ->where('volume_id', $session_cart->volume_id)
+                ->where('attribute_id', $session_cart->attribute_id)->first()) {
+                    $user_cart->quantity += $session_cart->quantity;
+
+                    $user_cart->save();
+
+                    $session_cart->delete();
+                } else {
+                    $session_cart->update([
+                        'user_id' => $user->id,
+                    ]);
+                }
             }
-            else {
-                $session_cart->update([
-                    'user_id' => $user->id,
-                ]);
+
+            foreach ($wish_ids as $id) {
+                $session_wish = Wishlist::find($id['id']);
+
+                if ($user_wish = Wishlist::where('user_id', $user->id)
+                ->where('volume_id', $session_wish->volume_id)
+                    ->where('attribute_id', $session_wish->attribute_id)->first()
+                ) {
+                    $user_wish->quantity += $session_wish->quantity;
+
+                    $user_wish->save();
+
+                    $session_wish->delete();
+                } else {
+                    $session_wish->update([
+                        'user_id' => $user->id,
+                    ]);
+                }
             }
+
+            Session::put('cart_quantity', (new CartService(new Cart()))->getItemValue());
+            Session::put('wish_quantity', (new WishlistService(new Wishlist()))->getItemValue());
+            Session::put('cart_price', (new CartService(new Cart()))->getAuthCartAmount());
         }
-
-        foreach($wish_ids as $id)
-        {
-            $session_wish = Wishlist::find($id['id']);
-
-            if($user_wish = Wishlist::where('user_id', $user->id)
-                ->where('volume_id',$session_wish->volume_id)
-                ->where('attribute_id',$session_wish->attribute_id)->first())
-            {
-                $user_wish->quantity += $session_wish->quantity;
-
-                $user_wish->save();
-
-                $session_wish->delete();
-            }
-            else {
-                $session_wish->update([
-                    'user_id' => $user->id,
-                ]);
-            }
-        }
-
-        Session::put('cart_quantity', (new CartService(new Cart()))->getItemValue());
-        Session::put('wish_quantity', (new WishlistService(new Wishlist()))->getItemValue());
-        Session::put('cart_price', (new CartService(new Cart()))->getAuthCartAmount());
     }
 }
